refactor(local-community): extract publication-validation helpers
Moves _checkPublicationValidity, _isPublicationAuthorPartOfRoles,
_respondWithErrorIfSignatureOfPublicationIsInvalid, and _isFlairInAllowedList
into a focused module. Splits the 428-line _checkPublicationValidity into a
short orchestrator (checkPublicationValidity) plus named sub-functions
(checkWireFormatAndCommunityAuthor, checkAuthorIdentity, checkParentAndPostState,
checkCommentPublication, checkVotePublication, checkCommentModerationPublication,
checkCommunityEditPublication, checkCommentEditPublication) without changing
the validation logic. Drops 'private' from _publishFailedChallengeVerification
so respondWithErrorIfSignatureOfPublicationIsInvalid can call it until
challenges.ts is extracted next.
